Mroizotelegrambot-main/Mroizotelegrambot-main/requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
praw==7.7.1
schedule==1.2.1
nest-asyncio==1.5.8
gTTS==2.4.0
SpeechRecognition==3.10.0
googletrans-py==4.0.0
requests==2.31.0
aiohttp==3.9.1
urllib3>=2.0.0
tornado==6.3.3
translate==3.6.1
backoff==2.2.1
pymongo==4.6.1
dnspython==2.7.0
motor==3.3.2
argostranslate